+++ package.json
@@ -9,8 +9,10 @@
     "repository": {
         "url": "https://github.com/fb55/css-what"
     },
-    "main": "lib/index.js",
+    "main": "lib/commonjs/index.js",
+    "module": "lib/es/index.js",
     "types": "lib/index.d.ts",
+    "sideEffects": false,
     "files": [
         "lib/**/*"
     ],
@@ -24,7 +26,7 @@
         "format:es": "npm run lint:es -- --fix",
         "format:prettier": "npm run prettier -- --write",
         "prettier": "prettier '**/*.{ts,md,json,yml}'",
-        "build": "tsc",
+        "build": "tsc && tsc -p tsconfig.es.json",
         "prepare": "npm run build"
     },
     "devDependencies": {
@@ -45,7 +47,10 @@
     },
     "license": "BSD-2-Clause",
     "jest": {
-        "preset": "ts-jest"
+        "preset": "ts-jest",
+        "roots": [
+            "src"
+        ]
     },
     "prettier": {
         "tabWidth": 4
